+++ lib/blog_snippets/articles/tuning_dd_block_size/dd_obs_test.sh
@@ -4,7 +4,13 @@
 set -e
 
 TEST_FILE=${1:-dd_obs_testfile}
+TEST_FILE_EXISTS=0
+if [ -e "$TEST_FILE" ]; then TEST_FILE_EXISTS=1; fi
 TEST_FILE_SIZE=134217728
+
+if [ $EUID -ne 0 ]; then
+  echo "NOTE: Kernel cache will not be cleared between tests without sudo. This will likely cause inaccurate results." 1>&2
+fi
 
 # Header
 PRINTF_FORMAT="%8s : %s\n"
@@ -21,13 +27,18 @@ do
     break
   fi
 
+  # Clear kernel cache to ensure more accurate test
+  [ $EUID -eq 0 ] && [ -e /proc/sys/vm/drop_caches ] && echo 3 > /proc/sys/vm/drop_caches
+
   # Create a test file with the specified block size
-  DD_RESULT=$(dd if=/dev/zero of=$TEST_FILE bs=$BLOCK_SIZE count=$COUNT 2>&1 1>/dev/null)
+  DD_RESULT=$(dd if=/dev/zero of=$TEST_FILE bs=$BLOCK_SIZE count=$COUNT conv=fsync 2>&1 1>/dev/null)
 
   # Extract the transfer rate from dd's STDERR output
   TRANSFER_RATE=$(echo $DD_RESULT | \grep --only-matching -E '[0-9.]+ ([MGk]?B|bytes)/s(ec)?')
 
-  # Clean up the test file and output result
-  rm $TEST_FILE
+  # Clean up the test file if we created one
+  if [ $TEST_FILE_EXISTS -ne 0 ]; then rm $TEST_FILE; fi
+
+  # Output the result
   printf "$PRINTF_FORMAT" "$BLOCK_SIZE" "$TRANSFER_RATE"
 done
